Consolidate Framework package
Move anything that has to do with framework in one folder
