--- shared/common.go
@@ -1,10 +1,0 @@
-package shared
-
-import "net/http"
-
-// FileProvider fetches files by name
-type FileProvider interface {
-	List() []string
-	String(name string) string
-	Open(name string) (http.File, error)
-}
